Squashed commit of the following:
commit 8ac411a3e20018fce3fa2b974dbd7b269617683b
Author: xsddxr909 <42365232+xsddxr909@users.noreply.github.com>
Date:   Sat Jun 13 18:30:16 2020 +0800

    Create 11

+++ batDemo/Assets/Scripts/Manager/PatchManager.cs
@@ -180,6 +180,7 @@ public class PatchManager : MonoSingleton<PatchManager>
                 signedFileName = fileInfo.signedFileName;
             }else{
                DebugLog.LogError("GetSignedFileLocalURL LogError"+fileName+" manifest"+manifest.fileInfoDict.Count);
+               return "";
             }
         }
         else
@@ -468,7 +469,7 @@ public class PatchManager : MonoSingleton<PatchManager>
         //顺便删除 旧版本的
         if(!AssetBundleFileInfo.IsConfigFile(fileInfo.fileName)){
              string  url= GetSignedFileLocalURL(fileInfo.fileName);
-             if(!url.Contains(_streamingAssetsURL)){
+             if(url!=""&&!url.Contains(_streamingAssetsURL)){
                  DebugLog.Log("DeleteFileSafely :"+url);
                 FileUtil.DeleteFileSafely(url);
              }
